# [WIP] Improve crypto performance