### Benchmark:
Benchmark run against a packet about 670 bytes long.

**Current code:**

- Time to encrypt an outgoing packet: 6.464us
- Memory allocations to encrypt an outgoing packet: 6320B
- Time to decrypt an incoming packet: 6.524us
- Memory allocations to decrypt an incoming packet: 5208B

**This branch:**

- Time to encrypt an outgoing packet: 4.054us
- Memory allocations to encrypt an outgoing packet: 848B +size of packet
- Time to decrypt an incoming packet: 3.688us
- Memory allocations to decrypt an incoming packet: 1008B + size of packet

Unfortunately it seems that about 2/3rds of the CPU time is spent computing the HMAC. I don't think we can do anything about that at the moment.

See also: #552