@yaakov-h 2018-06-03
`DebugLog`'s code coverage bounces up and down of it's own volition. #547.